DP-798 Support Python 3.6 by removing dataclasses
Dataclasses were introduced in Python 3.7. To be able to support
Python 3.6 in the SDK (as well as in Origo CLI later), the usage of
dataclasses is removed. They can be restored when we drop support for
Python 3.6, probably sometime after it has reached its end-of-life at
2021-12-23.

Seven dataclasses were defined in the codebase:

  TokenProvider
  ClientCredentialsProvider(TokenProvider)
  TokenServiceProvider(TokenProvider)

  PipelineBase
  Pipeline(PipelineBase)
  PipelineInput(PipelineBase)
  Schema(PipelineBase)

The three first classes, TokenProvider, ClientCredentialsProvider, and
TokenServiceProvider, are trivially rewritten to normal classes with an
__init__ method similar to what a dataclass would generate. These
changes should be simple to revert when dataclass usage is reintroduced
again by deleting the custom __init__ methods.

The remaining four classes, PipelineBase, Pipeline, PipelineInput, and
Schema, didn't appear to actually use any dataclass-specific
functionality. They defined their own __init__ and __repr__ methods,
making the ones generated by @dataclass moot. These classes can probably
be rewritten a bit to better leverage the advantages of dataclasses when
dataclass usage is reintroduced.