Add `Dataset.auto_create_edition`
Add a new method `Dataset.auto_create_edition` for creating a new
edition with an automatic name based on the current time.

Based on a similarly named method in `okdata-cli`, but extracted here
because it's useful in other cases in addition to the CLI. The CLI may
in turn be changed to use this new method.